+++ config/cst.py
@@ -1,12 +1,12 @@
 from enum import Enum
 
 SHORT_VERSION = "0.0.12"
-REV_VERSION = "0"
+REV_VERSION = "2"
 VERSION_DEV_PHASE = "alpha"
 VERSION = "{0}-{1}".format(SHORT_VERSION, VERSION_DEV_PHASE)
 LONG_VERSION = "{0}_{1}-{2}".format(SHORT_VERSION, REV_VERSION, VERSION_DEV_PHASE)
 
-ORIGIN_URL = "https://github.com/Trading-Bot/CryptoBot.git"
+ORIGIN_URL = "https://github.com/Drakkar-Software/OctoBot.git"
 
 MSECONDS_TO_SECONDS = 1000
 MINUTE_TO_SECONDS = 60
@@ -63,8 +63,8 @@ CONFIG_PORTFOLIO_TOTAL = "total"
 # Notification
 CONFIG_NOTIFICATION_INSTANCE = "notifier"
 CONFIG_CATEGORY_NOTIFICATION = "notification"
-NOTIFICATION_STARTING_MESSAGE = "CryptoBot v{0} starting...".format(LONG_VERSION)
-NOTIFICATION_STOPPING_MESSAGE = "CryptoBot v{0} stopping...".format(LONG_VERSION)
+NOTIFICATION_STARTING_MESSAGE = "OctoBot v{0} starting...".format(LONG_VERSION)
+NOTIFICATION_STOPPING_MESSAGE = "OctoBot v{0} stopping...".format(LONG_VERSION)
 
 # DEBUG options
 CONFIG_DEBUG_OPTION_PERF = "performance_analyser"
@@ -131,19 +131,19 @@ CONFIG_INTERFACES = "interfaces"
 CONFIG_INTERFACES_WEB = "web"
 CONFIG_INTERFACES_TELEGRAM = "telegram"
 
-# Packages
+# Tentacles (packages)
 GITHUB = "github"
 GITHUB_RAW_CONTENT_URL = "https://raw.githubusercontent.com"
 GITHUB_BASE_URL = "https://github.com"
-PACKAGES_PUBLIC_REPOSITORY = "Trading-Bot/CryptoBot-Packages"
-PACKAGES_PUBLIC_LIST = "packages_list.json"
-PACKAGES_DEFAULT_BRANCH = "master"
+TENTACLES_PUBLIC_REPOSITORY = "Drakkar-Software/OctoBot-Tentacles"
+TENTACLES_PUBLIC_LIST = "tentacles_list.json"
+TENTACLES_DEFAULT_BRANCH = "master"
 EVALUATOR_DEFAULT_FOLDER = "Default"
 EVALUATOR_ADVANCED_FOLDER = "Advanced"
-CONFIG_PACKAGES_KEY = "packages"
-PACKAGE_DESCRIPTION = "package_description"
-PACKAGE_DESCRIPTION_LOCALISATION = "localisation"
-PACKAGE_DESCRIPTION_IS_URL = "is_url"
+CONFIG_TENTACLES_KEY = "tentacles"
+TENTACLE_DESCRIPTION = "tentacle_description"
+TENTACLE_DESCRIPTION_LOCALISATION = "localisation"
+TENTACLE_DESCRIPTION_IS_URL = "is_url"
 
 
 class EvaluatorMatrixTypes(Enum):
@@ -246,7 +246,7 @@ class TradeOrderSide(Enum):
 
 
 class OrderStatus(Enum):
-    FILLED = "filled"
+    FILLED = "closed"
     OPEN = "open"
     PARTIALLY_FILLED = "partially_filled"
     CANCELED = "canceled"