+++ interfaces/trading_util.py
@@ -2,12 +2,16 @@ from interfaces import get_bot
 from trading.trader.portfolio import Portfolio
 
 
+def get_traders():
+    return [trader for trader in get_bot().get_exchange_traders().values()] + \
+           [trader for trader in get_bot().get_exchange_trader_simulators().values()]
+
+
 def get_portfolio_current_value():
     simulated_value = 0
     real_value = 0
+    traders = get_traders()
 
-    traders = [trader for trader in get_bot().get_exchange_traders().values()] + \
-              [trader for trader in get_bot().get_exchange_trader_simulators().values()]
     for trader in traders:
         trade_manager = trader.get_trades_manager()
 
@@ -28,9 +32,8 @@ def get_portfolio_current_value():
 def get_open_orders():
     simulated_open_orders = []
     real_open_orders = []
+    traders = get_traders()
 
-    traders = [trader for trader in get_bot().get_exchange_traders().values()] + \
-              [trader for trader in get_bot().get_exchange_trader_simulators().values()]
     for trader in traders:
         if trader.get_simulate():
             simulated_open_orders.append(trader.get_open_orders())
@@ -40,9 +43,33 @@ def get_open_orders():
     return real_open_orders, simulated_open_orders
 
 
+def cancel_all_open_orders():
+    for trader in get_traders():
+        trader.cancel_all_open_orders()
+
+
+def set_enable_trading(enable):
+    for trader in get_traders():
+        trader.set_enabled(enable)
+
+
+def get_trades_history():
+    simulated_trades_history = []
+    real_trades_history = []
+    traders = get_traders()
+
+    for trader in traders:
+        if trader.get_simulate():
+            simulated_trades_history.append(trader.get_trades_manager().get_trade_history())
+        else:
+            real_trades_history.append(trader.get_trades_manager().get_trade_history())
+
+    return real_trades_history, simulated_trades_history
+
+
 def set_risk(risk):
-    traders = [trader for trader in get_bot().get_exchange_traders().values()] + \
-              [trader for trader in get_bot().get_exchange_trader_simulators().values()]
+    traders = get_traders()
+
     for trader in traders:
         trader.set_risk(risk)
 
@@ -54,9 +81,10 @@ def get_risk():
 def get_global_profitability():
     simulated_global_profitability = 0
     real_global_profitability = 0
+    traders = get_traders()
+    simulated_full_origin_value = 0
+    real_full_origin_value = 0
 
-    traders = [trader for trader in get_bot().get_exchange_traders().values()] + \
-              [trader for trader in get_bot().get_exchange_trader_simulators().values()]
     for trader in traders:
         trade_manager = trader.get_trades_manager()
 
@@ -64,19 +92,26 @@ def get_global_profitability():
         current_value, _, _ = trade_manager.get_profitability()
 
         if trader.get_simulate():
+            simulated_full_origin_value += trade_manager.get_portfolio_origin_value()
             simulated_global_profitability += current_value
         else:
+            real_full_origin_value += trade_manager.get_portfolio_origin_value()
             real_global_profitability += current_value
 
-    return real_global_profitability, simulated_global_profitability
+    simulated_percent_profitability = simulated_global_profitability * 100 / simulated_full_origin_value \
+        if simulated_full_origin_value > 0 else 0
+    real_percent_profitability = real_global_profitability * 100 / real_full_origin_value \
+        if real_full_origin_value > 0 else 0
+
+    return real_global_profitability, simulated_global_profitability, \
+           real_percent_profitability, simulated_percent_profitability
 
 
 def get_portfolios():
     simulated_portfolios = []
     real_portfolios = []
+    traders = get_traders()
 
-    traders = [trader for trader in get_bot().get_exchange_traders().values()] + \
-              [trader for trader in get_bot().get_exchange_trader_simulators().values()]
     for trader in traders:
         if trader.get_simulate():
             simulated_portfolios.append(trader.get_portfolio().get_portfolio())
@@ -133,8 +168,8 @@ def get_trades_by_times_and_prices():
     real_trades_times = []
     real_trades_prices = []
 
-    traders = [trader for trader in get_bot().get_exchange_traders().values()] + \
-              [trader for trader in get_bot().get_exchange_trader_simulators().values()]
+    traders = get_traders()
+
     for trader in traders:
         for trade in trader.get_trades_manager().get_trade_history():
             if trader.get_simulate():