+++ tests/unit_tests/trading_tests/test_trades_manager.py
@@ -13,7 +13,7 @@ class TestTradesManager:
         config = load_test_config()
         exchange_manager = ExchangeManager(config, ccxt.binance, is_simulated=True)
         exchange_inst = exchange_manager.get_exchange()
-        trader_inst = Trader(config, exchange_inst)
+        trader_inst = Trader(config, exchange_inst, 1)
         trades_manager_inst = trader_inst.get_trades_manager()
         return config, exchange_inst, trader_inst, trades_manager_inst