+++ trading/trader/trades_manager.py
@@ -45,6 +45,9 @@ class TradesManager:
     def get_reference(self):
         return self.reference_market
 
+    def get_trade_history(self):
+        return self.trade_history
+
     def add_new_trade_in_history(self, trade):
         if trade not in self.trade_history:
             self.trade_history.append(trade)
@@ -97,9 +100,6 @@ class TradesManager:
         for trade in self.trade_history:
             self.trades_value += self._evaluate_value(trade.get_currency(), trade.get_quantity())
         return self.trades_value
-
-    def get_trade_history(self):
-        return self.trade_history
 
     def _update_portfolio_current_value(self):
         with self.portfolio as pf: